refactor(settings): split SettingsModule into AppearanceModule and NostrModule
SettingsModule exceeded the detekt TooManyFunctions threshold (12/11)
after merging ClientTagModule. Split by concern: UI appearance settings
(display, theme, language) and Nostr protocol settings (bootstrap relays,
client tag).
