+++ app/src/main/java/io/github/omochice/pinosu/feature/settings/di/AppearanceModule.kt
@@ -4,24 +4,14 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import io.github.omochice.pinosu.core.nip.nip89.ClientTagRepository
-import io.github.omochice.pinosu.feature.settings.data.repository.LocalClientTagRepository
 import io.github.omochice.pinosu.feature.settings.data.repository.LocalSettingsRepository
 import io.github.omochice.pinosu.feature.settings.domain.repository.SettingsRepository
-import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveBootstrapRelaysUseCase
-import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveBootstrapRelaysUseCaseImpl
-import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveClientTagEnabledUseCase
-import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveClientTagEnabledUseCaseImpl
 import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveDisplayModeUseCase
 import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveDisplayModeUseCaseImpl
 import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveLanguageModeUseCase
 import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveLanguageModeUseCaseImpl
 import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveThemeModeUseCase
 import io.github.omochice.pinosu.feature.settings.domain.usecase.ObserveThemeModeUseCaseImpl
-import io.github.omochice.pinosu.feature.settings.domain.usecase.SetBootstrapRelaysUseCase
-import io.github.omochice.pinosu.feature.settings.domain.usecase.SetBootstrapRelaysUseCaseImpl
-import io.github.omochice.pinosu.feature.settings.domain.usecase.SetClientTagEnabledUseCase
-import io.github.omochice.pinosu.feature.settings.domain.usecase.SetClientTagEnabledUseCaseImpl
 import io.github.omochice.pinosu.feature.settings.domain.usecase.SetDisplayModeUseCase
 import io.github.omochice.pinosu.feature.settings.domain.usecase.SetDisplayModeUseCaseImpl
 import io.github.omochice.pinosu.feature.settings.domain.usecase.SetLanguageModeUseCase
@@ -30,13 +20,13 @@ import io.github.omochice.pinosu.feature.settings.domain.usecase.SetThemeModeUse
 import io.github.omochice.pinosu.feature.settings.domain.usecase.SetThemeModeUseCaseImpl
 
 /**
- * Hilt module for Settings feature dependency injection.
+ * Hilt module for appearance-related settings.
  *
- * Provides Settings-related repositories and use cases.
+ * Provides repository, display mode, theme, and language dependencies.
  */
 @Module
 @InstallIn(SingletonComponent::class)
-object SettingsModule {
+object AppearanceModule {
 
   @Provides fun provideSettingsRepository(impl: LocalSettingsRepository): SettingsRepository = impl
 
@@ -59,26 +49,5 @@ object SettingsModule {
 
   @Provides
   fun provideObserveLanguageMode(impl: ObserveLanguageModeUseCaseImpl): ObserveLanguageModeUseCase =
-      impl
-
-  @Provides
-  fun provideSetBootstrapRelays(impl: SetBootstrapRelaysUseCaseImpl): SetBootstrapRelaysUseCase =
-      impl
-
-  @Provides
-  fun provideObserveBootstrapRelays(
-      impl: ObserveBootstrapRelaysUseCaseImpl
-  ): ObserveBootstrapRelaysUseCase = impl
-
-  @Provides
-  fun provideClientTagRepository(impl: LocalClientTagRepository): ClientTagRepository = impl
-
-  @Provides
-  fun provideObserveClientTagEnabled(
-      impl: ObserveClientTagEnabledUseCaseImpl
-  ): ObserveClientTagEnabledUseCase = impl
-
-  @Provides
-  fun provideSetClientTagEnabled(impl: SetClientTagEnabledUseCaseImpl): SetClientTagEnabledUseCase =
       impl
 }
